+++ Changelog.md
@@ -35,6 +35,7 @@
 - Added the polymorphic `source` relation/fields to the `Adjustment` model
 - Added the `getSource` and `setSource` methods to the `Adjustment` model
 - Changed the `setAdjustable` method to return the itself (was `void` before)
+- Changed (fixed, in fact) the `SimpleTax`/`SimpleTaxDeduction` adjusters to properly calculate the tax amount when it is included in the price
 - Changed the default-generated video hash to a base 58 ULID instead of nano id
 - Changed the minimum Laravel version requirements to v10.48, v11.44 and v12.2 respectively